Add triangle shape
This adds a small amount of vertex data to draw a triangle. It shows
off how vertex shaders work, but not really fragment shaders.

The bridging header allows the types declared in MetalTypes work in
both the metal shader code and swift.

This all converted to swift from this guide:
https://developer.apple.com/documentation/metal/hello_triangle