Add buffer of verticies
This adds a buffer of vertex data with many (2250) verticies. They
display in a large grid of quads.

https://developer.apple.com/documentation/metal/fundamental_lessons/basi
c_buffers

I had a pretty annoying bug for a while where it would draw most of the
quads, but then somewhere around the 350th quad, it would stop drawing
each vertex where it was supposed to be and instead draw only to the
middle.

This was caused by my metal shader specifiying the vertices variable in
the constant address space instead of the device address space . I
don't totally understand why that caused the bug, but from the
documentation:

>If multiple executing instances of a graphics or kernel function are
accessing the buffer using an index such as the vertex ID, fragment
coordinate, or the thread position in grid, you must allocate the
buffer in the device address space.

I am definitely accessing the array of verticies with a vertex ID.
Apparently this doesn't stop working until somewhere in the thousands
of verticies.