Display image as a texture
Some high level changes:

* The metal data type Vector removed color and added textureCoordinate
* There's also a new TextureIndexBaseColor type
* Images are loaded the a new TGAImage class
* The metal fragment shader got more complex (and renamed)

The TGAImage class is pretty interesting. First it loads in a file,
hopefully a TGA image. It then confirms various aspects of the file
(extension, type, color map, offsets, bits per pixel, etc). The texture
data needs to be in a very specific format so we need to ensure that we
can read it.

Then, if the image is 24 bits per pixel (8 bits each of blue, green,
and red) we need to convert it to 32 bits (blue, green, red, alpha).

Of note also, the TGAHeader properties are not in the same order as the
header itself so that the memory could be packed. Later, when I
reference the size of the header, it will return the correct size (18
instead of 20 bytes) because all of the properties are aligned.

This webpage was very helpful in informing me about the TGA header
format: http://paulbourke.net/dataformats/tga/